LumenAi/server/requirements.txt
fastapi
uvicorn
supabase
python-multipart
markitdown[pdf]

python-dotenv
google-genai
openai-whisper
ollama
httpx
